Renamed aux.js to util.js for WIN compatibility
