+++ test/6-wyvern-exchange-matching-erc1155.js
@@ -11,7 +11,7 @@ const Web3 = require('web3')
 const provider = new Web3.providers.HttpProvider('http://localhost:8545')
 const web3 = new Web3(provider)
 
-const {wrap,ZERO_BYTES32,CHAIN_ID,NULL_SIG,assertIsRejected} = require('./aux')
+const {wrap,ZERO_BYTES32,CHAIN_ID,NULL_SIG,assertIsRejected} = require('./util')
 
 contract('WyvernExchange', (accounts) =>
 	{
