+++ test/7-static-market-matching.js
@@ -12,7 +12,7 @@ const Web3 = require('web3')
 const provider = new Web3.providers.HttpProvider('http://localhost:8545')
 const web3 = new Web3(provider)
 
-const {wrap,ZERO_BYTES32,CHAIN_ID,assertIsRejected} = require('./aux')
+const {wrap,ZERO_BYTES32,CHAIN_ID,assertIsRejected} = require('./util')
 
 contract('WyvernExchange', (accounts) =>
 	{
